fix air visual, amartha and igrow (#75)
* fix air visual, amartha and igrow

* re-add airvisual death and recovered case

+++ src/main/scala/com/seanmcapp/external/IGrowClient.scala
@@ -2,7 +2,7 @@ package com.seanmcapp.external
 
 case class IgrowResponse(data: Seq[IgrowData])
 
-case class IgrowData(name: String, price: Long, stock: Int, `return`: String, contract_period: Int)
+case class IgrowData(name: String, price: Long, stock: Int, `return`: String, contract_period: Double)
 
 class IGrowClient(http: HttpRequestClient) {
